backend/requirements.txt
fastapi[standard]
google-genai==1.64.0
pillow==12.1.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose==3.5.0
supabase